refactor, bug : Refactor shelfIndex, searching, bug in Books.js
Refactor how the index was stored for the shelfIndex to allow easier retrieval. Refactor search to show a no results text when
search does not return books. Books had an error display the thumbnail when imageLinks.thumbnail was not in the json.
Search and Bookshelf both displayed books in the same manner, so refactored and added a BooksGrid component.